# precompute distances for faster merging